apply some suggested changes
